Pass arguments to condition procs based on arity
Fixes an issue where a lambda was expecting 1 argument but the call was
passing in 2 arguments (object and name). This resulted in an array
being passed as the (single) argument which broke the interface.

According to the documentation a single argument was expected:

```
node(:coolness, :if => lambda { |m| m.coolness > 5 }) do |m|
  m.coolness
end
```

Now we check the arity of the passed in lambda or Proc and based on the
arity we pass in the expected arguments. This makes the call compatible
with the previous (expected) behaviour and the documentation.