Fix failing Rabl::Engine tests
The order of the objects in the rendered JSON is not guaranteed.
Sometimes the tests would fail depending on how the JSON was rendered as
a string. Fix by using JSON parsing to compare the objects.

In same (simple) cases JSON.parse is unnecessary. I added it to all
assertions for consistency.